Added murmur3 hash function

+++ include/types.h
@@ -20,7 +20,7 @@ typedef uint32_t gid_t;      // Used for group IDs.
 typedef uint32_t id_t;       // Used as a general identifier.
 typedef uint64_t ino_t;      // Used for file serial numbers.
 typedef uint32_t key_t;      // Used for XSI interprocess communication.
-typedef uint16_t mode_t;     // Used for some file attributes.
+typedef uint32_t mode_t;     // Used for some file attributes.
 typedef uint16_t nlink_t;    // Used for link counts.
 typedef int64_t off_t;       // Used for file sizes.
 typedef int32_t pid_t;       // Used for process IDs and process group IDs.
